refactor(upgrade): decompose execution god-file test module into submodules
Re-applied at the crates/homeboy-core path after #8389 extracted the
homeboy-core crate (moving src/core -> crates/homeboy-core/src).

upgrade/execution.rs was 2316 lines (god_file) with ~1073 lines of
production and a ~1240-line inline test module. Split the test module
into execution/tests/{part_a,part_b}.rs with shared helpers
(checkout_with_package_name, git, git_stdout, etc.) promoted to
execution/tests/mod.rs as pub(super).

execution.rs is now 1075 lines (production only). cargo check -p
homeboy-core --lib is clean. Clears the god_file finding.
