+++ crates/homeboy-core/src/upgrade/execution.rs
@@ -1072,1245 +1072,4 @@ fn parse_cli_build_identity_output(output: &str) -> Option<String> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn bound_captured_stream_retains_full_source_within_limit() {
-        let (text, capture) = bound_captured_stream(b"boom", 1024);
-
-        assert_eq!(text, "boom");
-        assert_eq!(capture.limit_bytes, 1024);
-        assert_eq!(capture.seen_bytes, 4);
-        assert_eq!(capture.retained_bytes, 4);
-        assert!(!capture.truncated);
-    }
-
-    #[test]
-    fn bound_captured_stream_keeps_trailing_tail_when_truncated() {
-        let source = vec![b'x'; 16];
-        let (text, capture) = bound_captured_stream(&source, 4);
-
-        assert_eq!(text, "xxxx");
-        assert_eq!(capture.limit_bytes, 4);
-        assert_eq!(capture.retained_bytes, 4);
-        assert_eq!(capture.seen_bytes, 16);
-        assert!(capture.truncated);
-    }
-
-    #[test]
-    fn bound_captured_stream_retains_most_relevant_tail() {
-        let source = b"head-noise-TAIL".to_vec();
-        let (text, capture) = bound_captured_stream(&source, 4);
-
-        assert_eq!(text, "TAIL");
-        assert!(capture.truncated);
-    }
-
-    #[test]
-    fn annotate_truncation_notes_dropped_bytes() {
-        let capture = StreamCaptureMetadata {
-            limit_bytes: 4,
-            seen_bytes: 16,
-            retained_bytes: 4,
-            truncated: true,
-        };
-
-        let annotated = annotate_truncation("TAIL", &capture);
-
-        assert!(annotated.starts_with("TAIL"));
-        assert!(annotated.contains("output truncated"));
-        assert!(annotated.contains("retained 4 of 16 bytes"));
-    }
-
-    #[test]
-    fn annotate_truncation_leaves_untruncated_detail_unchanged() {
-        let capture = StreamCaptureMetadata {
-            limit_bytes: 1024,
-            seen_bytes: 4,
-            retained_bytes: 4,
-            truncated: false,
-        };
-
-        assert_eq!(annotate_truncation("boom", &capture), "boom");
-    }
-
-    #[test]
-    fn parses_homeboy_version_output() {
-        assert_eq!(
-            parse_cli_version_output("homeboy 0.158.0").as_deref(),
-            Some("0.158.0")
-        );
-    }
-
-    #[test]
-    fn command_output_with_timeout_captures_child_output() {
-        let mut command = Command::new("sh");
-        command.args(["-c", "printf 'homeboy 0.247.5'; printf 'warn' >&2"]);
-
-        let output = command_output_with_timeout(&mut command, Duration::from_secs(5))
-            .expect("command output");
-
-        assert!(output.status.success());
-        assert_eq!(String::from_utf8_lossy(&output.stdout), "homeboy 0.247.5");
-        assert_eq!(String::from_utf8_lossy(&output.stderr), "warn");
-    }
-
-    #[test]
-    fn source_upgrade_command_returns_after_same_binary_success() {
-        let workspace = tempfile::tempdir().expect("workspace");
-
-        run_source_upgrade_command(
-            "printf 'built same-version binary\\n'",
-            workspace.path(),
-            Duration::from_secs(1),
-        )
-        .expect("source command completes");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn source_upgrade_timeout_terminates_the_entire_child_process_group() {
-        let workspace = tempfile::tempdir().expect("workspace");
-        let pid_file = workspace.path().join("child.pid");
-        let command = format!("sleep 30 & echo $! > {}; wait", shell_quote_path(&pid_file));
-
-        let err = run_source_upgrade_command(&command, workspace.path(), Duration::from_millis(50))
-            .expect_err("long-running source command times out");
-        assert!(
-            err.details.to_string().to_lowercase().contains("timed out"),
-            "unexpected timeout error: {err:?}"
-        );
-
-        let child_pid = std::fs::read_to_string(&pid_file)
-            .expect("background child pid")
-            .trim()
-            .parse::<i32>()
-            .expect("numeric pid");
-        // The shell is the process-group leader and timeout termination must
-        // remove its background child as well as reap the direct child.
-        for _ in 0..40 {
-            if unsafe { libc::kill(child_pid, 0) } != 0 {
-                return;
-            }
-            std::thread::sleep(Duration::from_millis(25));
-        }
-        panic!("source-upgrade child must not be orphaned");
-    }
-
-    #[test]
-    fn test_execute_upgrade() {
-        assert_eq!(
-            parse_cli_version_output("homeboy 0.158.0").as_deref(),
-            Some("0.158.0")
-        );
-        assert!(!upgrade_verification_result(
-            InstallMethod::Source,
-            false,
-            "0.157.1",
-            Some("0.157.1"),
-            Some("commit old, dirty=false"),
-            Some("commit new, dirty=false"),
-        ));
-    }
-
-    #[test]
-    fn test_upgrade_verification_result() {
-        assert!(upgrade_verification_result(
-            InstallMethod::Secondary,
-            false,
-            "0.157.1",
-            Some("0.158.0"),
-            None,
-            None,
-        ));
-        assert!(!upgrade_verification_result(
-            InstallMethod::Secondary,
-            false,
-            "0.157.1",
-            Some("0.157.1"),
-            Some("commit old, dirty=false"),
-            Some("commit new, dirty=false"),
-        ));
-        assert!(!upgrade_verification_result(
-            InstallMethod::Source,
-            true,
-            "0.157.1",
-            None,
-            Some("commit old, dirty=false"),
-            Some("commit new, dirty=false"),
-        ));
-    }
-
-    #[test]
-    fn forced_source_upgrade_rejects_unchanged_same_version_build_identity() {
-        assert!(!upgrade_verification_result(
-            InstallMethod::Source,
-            true,
-            "0.157.1",
-            Some("0.157.1"),
-            Some("commit same, dirty=false"),
-            Some("commit same, dirty=false"),
-        ));
-    }
-
-    #[test]
-    fn forced_secondary_upgrade_accepts_same_version_active_binary() {
-        assert!(upgrade_verification_result(
-            InstallMethod::Secondary,
-            true,
-            "0.157.1",
-            Some("0.157.1"),
-            Some("commit same, dirty=false"),
-            Some("commit same, dirty=false"),
-        ));
-    }
-
-    #[test]
-    fn verification_accepts_newer_active_binary() {
-        assert!(upgrade_verification_result(
-            InstallMethod::Secondary,
-            false,
-            "0.157.1",
-            Some("0.158.0"),
-            None,
-            None,
-        ));
-    }
-
-    #[test]
-    fn verification_rejects_missing_active_binary_version() {
-        assert!(!upgrade_verification_result(
-            InstallMethod::Source,
-            true,
-            "0.157.1",
-            None,
-            Some("commit old, dirty=false"),
-            Some("commit new, dirty=false"),
-        ));
-    }
-
-    #[test]
-    fn forced_source_upgrade_accepts_same_version_with_new_build_identity() {
-        assert!(upgrade_verification_result(
-            InstallMethod::Source,
-            true,
-            "0.157.1",
-            Some("0.157.1"),
-            Some("homeboy 0.157.1+old"),
-            Some("homeboy 0.157.1+new"),
-        ));
-    }
-
-    #[test]
-    fn source_install_byte_match_rejects_same_version_stale_binary() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let source = dir.path().join("source");
-        let built = source.join("target/release/homeboy");
-        let active = dir.path().join("bin/homeboy");
-        std::fs::create_dir_all(built.parent().expect("built parent")).expect("built parent dir");
-        std::fs::create_dir_all(active.parent().expect("active parent"))
-            .expect("active parent dir");
-        std::fs::write(&built, b"homeboy 0.281.20 with new source behavior")
-            .expect("write built binary");
-        std::fs::write(&active, b"homeboy 0.281.20 stale installed binary")
-            .expect("write active binary");
-
-        assert!(
-            !upgrade_verification_result(
-                InstallMethod::Source,
-                true,
-                "0.281.20",
-                Some("0.281.20"),
-                Some("homeboy 0.281.20"),
-                Some("homeboy 0.281.20"),
-            ),
-            "identity-only verification cannot prove same-version source replacement"
-        );
-        assert!(
-            !source_install_matches_binary_path(&built, &active).expect("compare binaries"),
-            "same-version stale active binary must not verify"
-        );
-
-        std::fs::copy(&built, &active).expect("install built binary");
-
-        assert!(
-            source_install_matches_binary_path(&built, &active).expect("compare binaries"),
-            "source upgrade only verifies after the active binary is the built artifact"
-        );
-    }
-
-    #[test]
-    fn source_built_binary_path_uses_source_local_target_by_default() {
-        let workspace = Path::new("/workspace/homeboy");
-
-        assert_eq!(
-            source_built_binary_path_for_target_dir(workspace, None),
-            workspace.join("target/release/homeboy")
-        );
-    }
-
-    #[test]
-    fn source_built_binary_path_uses_absolute_cargo_target_dir() {
-        let workspace = Path::new("/workspace/homeboy");
-        let target_dir = Path::new("/shared/cargo-target");
-
-        assert_eq!(
-            source_built_binary_path_for_target_dir(workspace, Some(target_dir.as_os_str())),
-            target_dir.join("release/homeboy")
-        );
-    }
-
-    #[test]
-    fn source_built_binary_path_resolves_relative_cargo_target_dir_from_workspace() {
-        let workspace = Path::new("/workspace/homeboy");
-
-        assert_eq!(
-            source_built_binary_path_for_target_dir(
-                workspace,
-                Some(std::ffi::OsStr::new("shared/target"))
-            ),
-            workspace.join("shared/target/release/homeboy")
-        );
-    }
-
-    #[test]
-    fn forced_source_upgrade_accepts_same_version_without_build_identity() {
-        assert!(upgrade_verification_result(
-            InstallMethod::Source,
-            true,
-            "0.157.1",
-            Some("0.157.1"),
-            None,
-            Some("homeboy 0.157.1+new"),
-        ));
-    }
-
-    #[test]
-    fn non_forced_upgrade_rejects_same_version_active_binary() {
-        assert!(!upgrade_verification_result(
-            InstallMethod::Source,
-            false,
-            "0.157.1",
-            Some("0.157.1"),
-            Some("homeboy 0.157.1+old"),
-            Some("homeboy 0.157.1+old"),
-        ));
-    }
-
-    #[test]
-    fn parses_homeboy_version_output_with_build_identity() {
-        let info = parse_cli_version_info("homeboy 0.158.0+abc123-dirty");
-
-        assert_eq!(info.version.as_deref(), Some("0.158.0"));
-        assert_eq!(
-            info.build_identity.as_deref(),
-            Some("homeboy 0.158.0+abc123-dirty")
-        );
-    }
-
-    #[test]
-    fn parses_plain_homeboy_version_output_as_build_identity() {
-        let info = parse_cli_version_info("homeboy 0.158.0");
-
-        assert_eq!(info.version.as_deref(), Some("0.158.0"));
-        assert_eq!(info.build_identity.as_deref(), Some("homeboy 0.158.0"));
-    }
-
-    #[test]
-    fn verify_retry_succeeds_after_transient_unreadable_binary() {
-        // Issue #3463: the swap succeeded but the first read-back of the new
-        // binary returns nothing (racing the just-replaced binary). A later
-        // attempt reports the upgraded version and verification must succeed.
-        let reads = std::cell::RefCell::new(vec![
-            None,
-            Some(ActiveBinaryInfo {
-                version: Some("0.220.3".to_string()),
-                build_identity: None,
-            }),
-        ]);
-        let mut sleeps = 0u32;
-
-        let (success, active) = verify_upgrade_with_retry(
-            InstallMethod::Binary,
-            false,
-            "0.220.0",
-            None,
-            5,
-            std::time::Duration::from_millis(0),
-            || reads.borrow_mut().remove(0),
-            |_| sleeps += 1,
-        );
-
-        assert!(success, "transient read-back failure should be retried");
-        assert_eq!(
-            active.and_then(|info| info.version).as_deref(),
-            Some("0.220.3")
-        );
-        assert_eq!(sleeps, 1, "should sleep once between the two attempts");
-    }
-
-    #[test]
-    fn verify_retry_succeeds_after_stale_old_version() {
-        // The read-back briefly reports the old version before the new binary
-        // is observable; the retry should pick up the upgraded version.
-        let reads = std::cell::RefCell::new(vec![
-            Some(ActiveBinaryInfo {
-                version: Some("0.220.0".to_string()),
-                build_identity: None,
-            }),
-            Some(ActiveBinaryInfo {
-                version: Some("0.220.3".to_string()),
-                build_identity: None,
-            }),
-        ]);
-
-        let (success, active) = verify_upgrade_with_retry(
-            InstallMethod::Binary,
-            false,
-            "0.220.0",
-            None,
-            5,
-            std::time::Duration::from_millis(0),
-            || reads.borrow_mut().remove(0),
-            |_| {},
-        );
-
-        assert!(success);
-        assert_eq!(
-            active.and_then(|info| info.version).as_deref(),
-            Some("0.220.3")
-        );
-    }
-
-    #[test]
-    fn verify_retry_first_attempt_success_does_not_sleep() {
-        let mut sleeps = 0u32;
-
-        let (success, active) = verify_upgrade_with_retry(
-            InstallMethod::Binary,
-            false,
-            "0.220.0",
-            None,
-            5,
-            std::time::Duration::from_millis(0),
-            || {
-                Some(ActiveBinaryInfo {
-                    version: Some("0.220.3".to_string()),
-                    build_identity: None,
-                })
-            },
-            |_| sleeps += 1,
-        );
-
-        assert!(success);
-        assert_eq!(
-            active.and_then(|info| info.version).as_deref(),
-            Some("0.220.3")
-        );
-        assert_eq!(sleeps, 0, "no retries needed when first read verifies");
-    }
-
-    #[test]
-    fn verify_retry_exhausts_attempts_when_never_readable() {
-        let mut reads = 0u32;
-        let mut sleeps = 0u32;
-
-        let (success, active) = verify_upgrade_with_retry(
-            InstallMethod::Binary,
-            false,
-            "0.220.0",
-            None,
-            3,
-            std::time::Duration::from_millis(0),
-            || {
-                reads += 1;
-                None
-            },
-            |_| sleeps += 1,
-        );
-
-        assert!(
-            !success,
-            "genuinely unverifiable upgrade still reports false"
-        );
-        assert!(active.is_none());
-        assert_eq!(reads, 3, "all attempts consumed");
-        assert_eq!(sleeps, 2, "sleeps between attempts but not after the last");
-    }
-
-    #[test]
-    fn verify_retry_reports_last_seen_version_on_exhaustion() {
-        // The new version never becomes observable, but a stale old-version
-        // read is retained so the caller can still surface a version string.
-        let (success, active) = verify_upgrade_with_retry(
-            InstallMethod::Binary,
-            false,
-            "0.220.0",
-            None,
-            2,
-            std::time::Duration::from_millis(0),
-            || {
-                Some(ActiveBinaryInfo {
-                    version: Some("0.220.0".to_string()),
-                    build_identity: None,
-                })
-            },
-            |_| {},
-        );
-
-        assert!(!success);
-        assert_eq!(
-            active.and_then(|info| info.version).as_deref(),
-            Some("0.220.0")
-        );
-    }
-
-    #[test]
-    fn test_verify_upgrade_with_retry() {
-        // Smoke test covering the happy path with a single immediate read.
-        let (success, active) = verify_upgrade_with_retry(
-            InstallMethod::Secondary,
-            false,
-            "0.157.1",
-            None,
-            1,
-            std::time::Duration::from_millis(0),
-            || {
-                Some(ActiveBinaryInfo {
-                    version: Some("0.158.0".to_string()),
-                    build_identity: None,
-                })
-            },
-            |_| {},
-        );
-
-        assert!(success);
-        assert_eq!(
-            active.and_then(|info| info.version).as_deref(),
-            Some("0.158.0")
-        );
-    }
-
-    #[test]
-    fn source_swap_failure_errors_when_active_binary_unchanged() {
-        // Issue #5772: the source upgrade command exited 0 but the read-back
-        // proves the active binary was not replaced — fail loudly instead of a
-        // soft `upgraded: false` "completed".
-        let err = source_swap_failure(
-            InstallMethod::Source,
-            false,
-            Some("0.247.5"),
-            Some("homeboy 0.247.5+old"),
-            Some(Path::new("/src/homeboy")),
-            Some(Path::new("/src/homeboy/target/release/homeboy")),
-            Some(Path::new("/active/homeboy")),
-        )
-        .expect("unverified source swap must surface an error");
-
-        assert!(err.message.contains("active binary was not replaced"));
-        assert!(err.message.contains("homeboy 0.247.5+old"));
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("--method source")));
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("Active binary path:")));
-        assert!(err.hints.iter().any(|hint| hint
-            .message
-            .contains("Built source binary: /src/homeboy/target/release/homeboy")));
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("Replacement target path:")));
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("Permissions:")));
-    }
-
-    #[test]
-    fn source_swap_failure_reports_version_when_no_build_identity() {
-        let err = source_swap_failure(
-            InstallMethod::Source,
-            false,
-            Some("0.247.5"),
-            None,
-            Some(Path::new("/src/homeboy")),
-            Some(Path::new("/src/homeboy/target/release/homeboy")),
-            None,
-        )
-        .expect("unverified source swap must surface an error");
-
-        assert!(err.message.contains("0.247.5"));
-    }
-
-    #[test]
-    fn source_swap_failure_reports_placeholder_when_version_unverifiable() {
-        let err = source_swap_failure(
-            InstallMethod::Source,
-            false,
-            None,
-            None,
-            Some(Path::new("/src/homeboy")),
-            Some(Path::new("/src/homeboy/target/release/homeboy")),
-            None,
-        )
-        .expect("unverified source swap must surface an error");
-
-        assert!(err.message.contains("an unverifiable version"));
-    }
-
-    #[test]
-    fn source_swap_failure_silent_on_verified_swap() {
-        assert!(
-            source_swap_failure(
-                InstallMethod::Source,
-                true,
-                Some("0.249.0"),
-                None,
-                Some(Path::new("/src/homeboy")),
-                Some(Path::new("/src/homeboy/target/release/homeboy")),
-                None,
-            )
-            .is_none(),
-            "a verified source swap is not a failure"
-        );
-    }
-
-    #[test]
-    fn source_swap_failure_ignores_non_source_methods() {
-        // Non-source methods keep their soft unverified reporting; only source
-        // (where the swap is part of the command's contract) fails loudly here.
-        assert!(source_swap_failure(
-            InstallMethod::Binary,
-            false,
-            Some("0.247.5"),
-            None,
-            Some(Path::new("/src/homeboy")),
-            Some(Path::new("/src/homeboy/target/release/homeboy")),
-            None,
-        )
-        .is_none());
-        assert!(source_swap_failure(
-            InstallMethod::Secondary,
-            false,
-            Some("0.247.5"),
-            None,
-            Some(Path::new("/src/homeboy")),
-            Some(Path::new("/src/homeboy/target/release/homeboy")),
-            None,
-        )
-        .is_none());
-    }
-
-    #[test]
-    fn source_swap_failure_diagnostics_use_replacement_target_path() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let source = dir.path().join("source");
-        let target = dir.path().join("configured-runner-homeboy");
-        std::fs::create_dir_all(source.join("target/release")).expect("source dirs");
-        std::fs::write(&target, "old").expect("target");
-
-        let err = source_swap_failure(
-            InstallMethod::Source,
-            false,
-            Some("0.255.8"),
-            Some("homeboy 0.255.8+old"),
-            Some(&source),
-            Some(&source.join("target/release/homeboy")),
-            Some(&target),
-        )
-        .expect("unverified source swap must surface an error");
-
-        assert!(err.hints.iter().any(|hint| hint
-            .message
-            .contains(&format!("Active binary path: {}", target.display()))));
-        assert!(err.hints.iter().any(|hint| hint
-            .message
-            .contains(&format!("Replacement target path: {}", target.display()))));
-        let first_recovery = err
-            .hints
-            .iter()
-            .find(|hint| hint.message.contains("Homeboy") || hint.message.contains("replacement"))
-            .map(|hint| hint.message.as_str())
-            .unwrap_or_default();
-        assert!(first_recovery.contains("Retry through just-built Homeboy"));
-    }
-
-    #[test]
-    fn source_swap_failure_diagnostics_include_paths_permissions_and_remediation() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let source = dir.path().join("source checkout");
-        let bin_dir = dir.path().join("bin dir");
-        std::fs::create_dir_all(source.join("target/release")).expect("source dirs");
-        std::fs::create_dir_all(&bin_dir).expect("bin dir");
-        let active = bin_dir.join("homeboy active");
-        let built = source.join("target/release/homeboy");
-        std::fs::write(&active, "old").expect("active");
-        std::fs::write(&built, "new").expect("built");
-
-        let diagnostics =
-            source_swap_failure_diagnostics_for_paths(Some(&source), Some(&built), Some(&active));
-
-        assert_eq!(diagnostics.active_path, active.display().to_string());
-        assert_eq!(diagnostics.built_binary_path, built.display().to_string());
-        assert_eq!(diagnostics.replacement_path, active.display().to_string());
-        assert!(diagnostics.permissions.contains("active="));
-        assert!(diagnostics.permissions.contains("parent="));
-        assert!(diagnostics.permissions.contains("writable="));
-        let expected_copy_command = format!(
-            "install -m 0755 '{}' '{}'",
-            built.display(),
-            active.display()
-        );
-        assert_eq!(
-            diagnostics.copy_command.as_deref(),
-            Some(expected_copy_command.as_str())
-        );
-        let built_command = diagnostics
-            .built_binary_command
-            .as_deref()
-            .expect("built command");
-        assert!(built_command.contains("upgrade --method source"));
-        assert!(built_command.contains("--source-path"));
-        assert!(built_command.contains(&source.display().to_string()));
-    }
-
-    #[test]
-    fn source_swap_failure_diagnostics_report_effective_cargo_target_binary() {
-        let source = Path::new("/src/homeboy");
-        let built = Path::new("/shared/cargo-target/release/homeboy");
-        let active = Path::new("/bin/homeboy");
-
-        let diagnostics =
-            source_swap_failure_diagnostics_for_paths(Some(source), Some(built), Some(active));
-
-        assert_eq!(diagnostics.built_binary_path, built.display().to_string());
-        assert!(diagnostics
-            .copy_command
-            .as_deref()
-            .expect("copy command")
-            .contains("'/shared/cargo-target/release/homeboy'"));
-    }
-
-    #[test]
-    fn install_source_built_binary_replaces_active_target() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let source = dir.path().join("source");
-        let target_dir = dir.path().join("bin");
-        std::fs::create_dir_all(source.join("target/release")).expect("source dirs");
-        std::fs::create_dir_all(&target_dir).expect("target dir");
-        let built = source.join("target/release/homeboy");
-        let active = target_dir.join("homeboy");
-        std::fs::write(&built, "new homeboy").expect("built binary");
-        std::fs::write(&active, "old homeboy").expect("active binary");
-
-        install_source_built_binary(&built, &active).expect("install source binary");
-
-        assert_eq!(
-            std::fs::read_to_string(&active).expect("active"),
-            "new homeboy"
-        );
-        assert!(binary_files_match(&built, &active).expect("files match"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn install_source_built_binary_sets_executable_mode() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = tempfile::tempdir().expect("tempdir");
-        let source = dir.path().join("source");
-        let target_dir = dir.path().join("bin");
-        std::fs::create_dir_all(source.join("target/release")).expect("source dirs");
-        std::fs::create_dir_all(&target_dir).expect("target dir");
-        let built = source.join("target/release/homeboy");
-        let active = target_dir.join("homeboy");
-        std::fs::write(&built, "new homeboy").expect("built binary");
-        std::fs::write(&active, "old homeboy").expect("active binary");
-
-        install_source_built_binary(&built, &active).expect("install source binary");
-
-        let mode = std::fs::metadata(&active)
-            .expect("active metadata")
-            .permissions()
-            .mode()
-            & 0o777;
-        assert_eq!(mode, 0o755);
-    }
-
-    #[test]
-    fn install_source_built_binary_reports_copy_failure() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let source = dir.path().join("source");
-        let target_dir = dir.path().join("bin");
-        std::fs::create_dir_all(&source).expect("source dir");
-        std::fs::create_dir_all(&target_dir).expect("target dir");
-        let active = target_dir.join("homeboy");
-        std::fs::write(&active, "old homeboy").expect("active binary");
-
-        let built = source.join("target/release/homeboy");
-        let err = install_source_built_binary(&built, &active).expect_err("missing build fails");
-
-        let details = err.details.to_string();
-        assert!(details.contains("target/release/homeboy"));
-        assert!(details.contains("install source-built binary"));
-    }
-
-    #[test]
-    fn shell_quote_handles_paths_with_single_quotes() {
-        assert_eq!(quote_path("/tmp/homeboy's/bin"), "'/tmp/homeboy'\\''s/bin'");
-    }
-
-    #[test]
-    fn test_resolve_source_workspace() {
-        let dir = checkout_with_package_name("homeboy");
-
-        let resolved = resolve_source_workspace(Some(dir.path())).expect("source checkout");
-
-        assert_eq!(resolved, dir.path());
-    }
-
-    #[test]
-    fn source_workspace_rejects_non_homeboy_checkout() {
-        let dir = checkout_with_package_name("other");
-
-        let err = resolve_source_workspace(Some(dir.path())).expect_err("invalid checkout");
-
-        assert!(err.message.contains("Homeboy source workspace"));
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("--source-path")));
-    }
-
-    #[test]
-    fn source_workspace_accepts_snapshot_without_git_metadata() {
-        let dir = source_workspace_with_package_name("homeboy");
-
-        let resolved = resolve_source_workspace(Some(dir.path())).expect("source snapshot");
-
-        assert_eq!(resolved, dir.path());
-    }
-
-    #[test]
-    fn source_workspace_resolves_from_nested_checkout_path() {
-        let dir = checkout_with_package_name("homeboy");
-        let nested = dir.path().join("src/core");
-        std::fs::create_dir_all(&nested).expect("nested dir");
-
-        let resolved = resolve_source_workspace(Some(&nested)).expect("source checkout");
-
-        assert_eq!(resolved, dir.path());
-    }
-
-    #[test]
-    fn source_upgrade_preparation_preserves_detached_checkout_identity() {
-        let remote = tempfile::tempdir().expect("remote tempdir");
-        git(
-            remote.path(),
-            &["init", "--bare", "--initial-branch", "main"],
-        );
-
-        let seed = source_workspace_with_package_name("homeboy");
-        git(seed.path(), &["init", "--initial-branch", "main"]);
-        git(seed.path(), &["config", "user.email", "test@example.com"]);
-        git(seed.path(), &["config", "user.name", "Homeboy Test"]);
-        git(seed.path(), &["add", "."]);
-        git(seed.path(), &["commit", "-m", "initial"]);
-        git(
-            seed.path(),
-            &[
-                "remote",
-                "add",
-                "origin",
-                &remote.path().display().to_string(),
-            ],
-        );
-        git(seed.path(), &["push", "-u", "origin", "main"]);
-
-        let checkout = tempfile::tempdir().expect("checkout tempdir");
-        std::fs::remove_dir(checkout.path()).expect("remove placeholder checkout dir");
-        git(
-            remote.path(),
-            &[
-                "clone",
-                &remote.path().display().to_string(),
-                &checkout.path().display().to_string(),
-            ],
-        );
-
-        std::fs::write(seed.path().join("src.txt"), "new source\n").expect("write source change");
-        git(seed.path(), &["add", "."]);
-        git(seed.path(), &["commit", "-m", "update source"]);
-        git(seed.path(), &["push", "origin", "main"]);
-
-        git(checkout.path(), &["switch", "--detach", "HEAD"]);
-        let stale_head = git_stdout(checkout.path(), &["rev-parse", "HEAD"]);
-
-        let source_revision = prepare_source_workspace_for_upgrade(checkout.path())
-            .expect("prepare detached checkout");
-
-        let prepared_head = git_stdout(checkout.path(), &["rev-parse", "HEAD"]);
-        assert_eq!(prepared_head, stale_head);
-        assert_eq!(source_revision.as_deref(), Some(stale_head.as_str()));
-        assert_eq!(
-            git_stdout(checkout.path(), &["branch", "--show-current"]),
-            ""
-        );
-    }
-
-    #[test]
-    fn source_upgrade_preparation_preserves_local_only_worktree_branch() {
-        let remote = tempfile::tempdir().expect("remote tempdir");
-        git(
-            remote.path(),
-            &["init", "--bare", "--initial-branch", "main"],
-        );
-
-        let seed = source_workspace_with_package_name("homeboy");
-        git(seed.path(), &["init", "--initial-branch", "main"]);
-        git(seed.path(), &["config", "user.email", "test@example.com"]);
-        git(seed.path(), &["config", "user.name", "Homeboy Test"]);
-        git(seed.path(), &["add", "."]);
-        git(seed.path(), &["commit", "-m", "initial"]);
-        git(
-            seed.path(),
-            &[
-                "remote",
-                "add",
-                "origin",
-                &remote.path().display().to_string(),
-            ],
-        );
-        git(seed.path(), &["push", "-u", "origin", "main"]);
-
-        let root = tempfile::tempdir().expect("root tempdir");
-        let main_checkout = root.path().join("main-checkout");
-        git(
-            root.path(),
-            &[
-                "clone",
-                &remote.path().display().to_string(),
-                &main_checkout.display().to_string(),
-            ],
-        );
-
-        let source_worktree = root.path().join("source-worktree");
-        git(
-            &main_checkout,
-            &[
-                "worktree",
-                "add",
-                "-b",
-                "feature-upgrade-source",
-                &source_worktree.display().to_string(),
-                "HEAD",
-            ],
-        );
-
-        std::fs::write(seed.path().join("src.txt"), "new source\n").expect("write source change");
-        git(seed.path(), &["add", "."]);
-        git(seed.path(), &["commit", "-m", "update source"]);
-        git(seed.path(), &["push", "origin", "main"]);
-
-        let switch_main = std::process::Command::new("git")
-            .arg("-C")
-            .arg(&source_worktree)
-            .args(["switch", "main"])
-            .output()
-            .expect("run git switch main");
-        assert!(
-            !switch_main.status.success(),
-            "test setup should reproduce branch ownership failure"
-        );
-
-        let source_head = git_stdout(&source_worktree, &["rev-parse", "HEAD"]);
-        let source_revision = prepare_source_workspace_for_upgrade(&source_worktree)
-            .expect("prepare source worktree");
-
-        assert_eq!(
-            git_stdout(&source_worktree, &["branch", "--show-current"]),
-            "feature-upgrade-source"
-        );
-        assert_eq!(
-            git_stdout(&source_worktree, &["rev-parse", "HEAD"]),
-            source_head
-        );
-        assert_eq!(source_revision.as_deref(), Some(source_head.as_str()));
-
-        let command = source_upgrade_command_for_prepared_workspace(
-            "git pull --ff-only\ncargo build --release",
-            &source_worktree,
-            true,
-        )
-        .expect("explicit source command");
-        assert!(command
-            .contains("Skipping git $homeboy_git_arg for explicitly selected source checkout"));
-    }
-
-    #[test]
-    fn source_upgrade_preparation_rejects_dirty_source_checkout() {
-        let remote = tempfile::tempdir().expect("remote tempdir");
-        git(
-            remote.path(),
-            &["init", "--bare", "--initial-branch", "main"],
-        );
-
-        let seed = source_workspace_with_package_name("homeboy");
-        git(seed.path(), &["init", "--initial-branch", "main"]);
-        git(seed.path(), &["config", "user.email", "test@example.com"]);
-        git(seed.path(), &["config", "user.name", "Homeboy Test"]);
-        git(seed.path(), &["add", "."]);
-        git(seed.path(), &["commit", "-m", "initial"]);
-        git(
-            seed.path(),
-            &[
-                "remote",
-                "add",
-                "origin",
-                &remote.path().display().to_string(),
-            ],
-        );
-        git(seed.path(), &["push", "-u", "origin", "main"]);
-        std::fs::write(seed.path().join("uncommitted.txt"), "dirty\n").expect("dirty file");
-
-        let err = prepare_source_workspace_for_upgrade(seed.path()).expect_err("dirty rejected");
-
-        assert!(err.message.contains("uncommitted changes"));
-    }
-
-    #[test]
-    fn source_upgrade_preparation_rejects_checkout_without_a_commit() {
-        let checkout = source_workspace_with_package_name("homeboy");
-        git(checkout.path(), &["init", "--initial-branch", "main"]);
-
-        let err =
-            source_workspace_revision(checkout.path()).expect_err("unverified checkout rejected");
-
-        assert!(err.message.contains("immutable commit"));
-    }
-
-    #[test]
-    fn source_upgrade_command_preserves_branch_checkout_command() {
-        let checkout = source_workspace_with_package_name("homeboy");
-        git(checkout.path(), &["init", "--initial-branch", "main"]);
-
-        let command = source_upgrade_command_for_prepared_workspace(
-            "git pull --ff-only\ncargo build --release",
-            checkout.path(),
-            false,
-        )
-        .expect("source command");
-
-        assert_eq!(command, "git pull --ff-only\ncargo build --release");
-    }
-
-    #[test]
-    fn explicit_source_upgrade_command_skips_pull_for_detached_checkout() {
-        let checkout = source_workspace_with_package_name("homeboy");
-        git(checkout.path(), &["init", "--initial-branch", "main"]);
-        git(
-            checkout.path(),
-            &["config", "user.email", "test@example.com"],
-        );
-        git(checkout.path(), &["config", "user.name", "Homeboy Test"]);
-        git(checkout.path(), &["add", "."]);
-        git(checkout.path(), &["commit", "-m", "initial"]);
-        git(checkout.path(), &["switch", "--detach", "HEAD"]);
-
-        let command = source_upgrade_command_for_prepared_workspace(
-            "git pull --ff-only\ncargo build --release",
-            checkout.path(),
-            true,
-        )
-        .expect("source command");
-
-        assert!(command
-            .contains("Skipping git $homeboy_git_arg for explicitly selected source checkout"));
-        assert!(command.contains("HOMEBOY_GIT_GUARD_DIR"));
-        assert!(command.contains("git pull --ff-only\ncargo build --release"));
-    }
-
-    #[test]
-    fn explicit_source_upgrade_command_skips_pull_for_local_only_branch() {
-        let checkout = source_workspace_with_package_name("homeboy");
-        git(checkout.path(), &["init", "--initial-branch", "local-only"]);
-        git(
-            checkout.path(),
-            &["config", "user.email", "test@example.com"],
-        );
-        git(checkout.path(), &["config", "user.name", "Homeboy Test"]);
-        git(checkout.path(), &["add", "."]);
-        git(checkout.path(), &["commit", "-m", "initial"]);
-
-        let command = source_upgrade_command_for_prepared_workspace(
-            "git pull --ff-only\ncargo build --release",
-            checkout.path(),
-            true,
-        )
-        .expect("source command");
-
-        assert!(command
-            .contains("Skipping git $homeboy_git_arg for explicitly selected source checkout"));
-        assert!(command.contains("git pull --ff-only\ncargo build --release"));
-    }
-
-    #[test]
-    fn explicit_source_upgrade_runs_build_phase_without_upstream_on_local_branch_or_detached_head()
-    {
-        for detached in [false, true] {
-            let checkout = source_workspace_with_package_name("homeboy");
-            git(checkout.path(), &["init", "--initial-branch", "local-only"]);
-            git(
-                checkout.path(),
-                &["config", "user.email", "test@example.com"],
-            );
-            git(checkout.path(), &["config", "user.name", "Homeboy Test"]);
-            git(checkout.path(), &["add", "."]);
-            git(checkout.path(), &["commit", "-m", "initial"]);
-            if detached {
-                git(checkout.path(), &["switch", "--detach", "HEAD"]);
-            }
-
-            let build_marker = checkout.path().join("build-phase");
-            let upgrade_command = format!(
-                "set -e\ngit fetch origin\ngit reset --hard origin/main\nsh -c 'git pull --ff-only'\nprintf built > {}",
-                shell_quote_path(&build_marker)
-            );
-            let command = source_upgrade_command_for_prepared_workspace(
-                &upgrade_command,
-                checkout.path(),
-                true,
-            )
-            .expect("explicit source command");
-
-            run_source_upgrade_command(&command, checkout.path(), Duration::from_secs(5))
-                .expect("guarded upgrade command reaches build phase");
-            assert_eq!(std::fs::read_to_string(build_marker).unwrap(), "built");
-        }
-    }
-
-    #[test]
-    fn source_upgrade_command_preserves_snapshot_command() {
-        let checkout = source_workspace_with_package_name("homeboy");
-
-        let command = source_upgrade_command_for_prepared_workspace(
-            "cargo build --release",
-            checkout.path(),
-            true,
-        )
-        .expect("source command");
-
-        assert_eq!(command, "cargo build --release");
-    }
-
-    #[test]
-    fn executable_workspace_only_resolves_target_build_paths() {
-        let path = Path::new("/repo/target/release/homeboy");
-        assert_eq!(
-            workspace_from_exe_path(path).as_deref(),
-            Some(Path::new("/repo"))
-        );
-
-        let installed = Path::new("/usr/local/bin/homeboy");
-        assert!(workspace_from_exe_path(installed).is_none());
-    }
-
-    #[test]
-    fn binary_404_upgrade_error_suggests_source_fallback() {
-        let err = upgrade_failure_error(
-            InstallMethod::Binary,
-            "curl: (22) The requested URL returned error: 404",
-        );
-
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("No release asset")));
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("--source-path")));
-    }
-
-    #[test]
-    fn missing_tool_upgrade_error_suggests_source_fallback() {
-        let err = upgrade_failure_error(
-            InstallMethod::Secondary,
-            &format!(
-                "sh: 1: {}: not found",
-                defaults::secondary_install_method_key()
-            ),
-        );
-
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("Required executable")));
-        assert!(err
-            .hints
-            .iter()
-            .any(|hint| hint.message.contains("--source-path")));
-    }
-
-    fn checkout_with_package_name(package_name: &str) -> tempfile::TempDir {
-        let dir = tempfile::tempdir().expect("tempdir");
-        std::fs::create_dir(dir.path().join(".git")).expect("git dir");
-        write_source_workspace_files(dir.path(), package_name);
-        dir
-    }
-
-    fn source_workspace_with_package_name(package_name: &str) -> tempfile::TempDir {
-        let dir = tempfile::tempdir().expect("tempdir");
-        write_source_workspace_files(dir.path(), package_name);
-        dir
-    }
-
-    fn write_source_workspace_files(path: &Path, package_name: &str) {
-        let manifest = serde_json::json!({ "id": package_name });
-        std::fs::write(path.join("homeboy.json"), manifest.to_string()).expect("manifest");
-        let package_manifest = ["Car", "go.toml"].concat();
-        std::fs::write(
-            path.join(package_manifest),
-            format!("[package]\nname = \"{package_name}\"\nversion = \"0.0.0\"\n"),
-        )
-        .expect("package manifest");
-    }
-
-    fn git(path: &Path, args: &[&str]) {
-        let output = std::process::Command::new("git")
-            .arg("-C")
-            .arg(path)
-            .args(args)
-            .output()
-            .expect("run git");
-        assert!(
-            output.status.success(),
-            "git {} failed: {}{}",
-            args.join(" "),
-            String::from_utf8_lossy(&output.stderr),
-            String::from_utf8_lossy(&output.stdout)
-        );
-    }
-
-    fn git_stdout(path: &Path, args: &[&str]) -> String {
-        let output = std::process::Command::new("git")
-            .arg("-C")
-            .arg(path)
-            .args(args)
-            .output()
-            .expect("run git");
-        assert!(
-            output.status.success(),
-            "git {} failed: {}{}",
-            args.join(" "),
-            String::from_utf8_lossy(&output.stderr),
-            String::from_utf8_lossy(&output.stdout)
-        );
-        String::from_utf8_lossy(&output.stdout).trim().to_string()
-    }
-}
+mod tests;
